pyproject.toml
[project]
name = "textedit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langchain-postgres>=0.0.15",
    "langgraph>=0.6.7",
    "pgvector<0.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
